- fixed some compiler errors in VS.NET - fixed problem with incorrect month number in zip file timestamps
git-svn-id: http://crashrpt.googlecode.com/svn@751 821a4644-a21a-0410-aec7-6199d47568cc

+++ branches/1.2/src/crashsender/src/CrashInfoReader.h
@@ -142,7 +142,7 @@ public:
   BOOL AddFilesToCrashDescriptionXML(std::vector<FileItem>);
 
   ErrorReportInfo& GetReport(int nIndex){ return m_Reports[nIndex]; }
-  int GetReportCount(){ return m_Reports.size(); }
+  int GetReportCount(){ return (int)m_Reports.size(); }
 
   BOOL GetLastRemindDate(SYSTEMTIME& LastDate);
   BOOL SetLastRemindDateToday();
